requirements.txt
langchain
openai
flake8
pytest
fastapi
uvicorn
python-dotenv
